requirements.txt
django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.3.1
pyotp==2.9.0
qrcode==8.2
Pillow==11.2.1
python-jose==3.3.0 